# unpeel group by 3 ways to enable vectorization