This rearranges the group key assignment in 2 steps - split the loop to compute the group ids from the loop which marks the existence of the groups in a block, and then specialises the group id calculation loop for cases where there are 2 groups and 3 groups respectively. This simplifies the group id calculation loop to the extent that C2 can vectorize it on JDK11. This speeds up the code up at least 5x compared to a loop which can handle any number of group by expressions in a microbenchmark small enough to examine the generated code:


```java
import org.openjdk.jmh.annotations.*;

import java.util.concurrent.ThreadLocalRandom;

public class GroupByBenchmark {


  @State(Scope.Benchmark)
  public static class OneState {

    @Param({"10", "100"})
    int cardinality0;

    @Param({"1024", "10000"})
    int numDocs;

    Generic generic;
    One one;

    int[] outGroupIds;
    int[] dictIds0;
    int[][] dictIds;

    @Setup(Level.Trial)
    public void setup() {
      one = new One(cardinality0);
      generic = new Generic(new int[]{cardinality0});
      outGroupIds = new int[numDocs];
      dictIds0 = new int[numDocs];
      for (int i = 0; i < numDocs; i++) {
        dictIds0[i] = ThreadLocalRandom.current().nextInt(cardinality0);
      }
      dictIds = new int[][] {dictIds0};
    }
  }

  @State(Scope.Benchmark)
  public static class TwoState {

    @Param({"10", "100"})
    int cardinality0;

    @Param({"3", "30"})
    int cardinality1;

    @Param("1024")
    int numDocs;

    Generic generic;
    Two two;

    int[] outGroupIds;
    int[] dictIds0;
    int[] dictIds1;
    int[][] dictIds;

    @Setup(Level.Trial)
    public void setup() {
      two = new Two(cardinality0, cardinality1);
      generic = new Generic(new int[]{cardinality0, cardinality1});
      outGroupIds = new int[numDocs];
      dictIds0 = new int[numDocs];
      dictIds1 = new int[numDocs];
      for (int i = 0; i < numDocs; i++) {
        dictIds0[i] = ThreadLocalRandom.current().nextInt(cardinality0);
        dictIds1[i] = ThreadLocalRandom.current().nextInt(cardinality1);
      }
      dictIds = new int[][] {dictIds0, dictIds1};
    }
  }

  @State(Scope.Benchmark)
  public static class ThreeState {

    @Param({"10", "100"})
    int cardinality0;

    @Param({"3", "30"})
    int cardinality1;

    @Param({"5", "50"})
    int cardinality2;

    @Param({"1024", "10000"})
    int numDocs;

    Generic generic;
    Three three;
    GenericFissured genericFissured;

    int[] outGroupIds;
    int[] dictIds0;
    int[] dictIds1;
    int[] dictIds2;
    int[][] dictIds;

    @Setup(Level.Trial)
    public void setup() {
      three = new Three(cardinality0, cardinality1, cardinality2);
      generic = new Generic(new int[]{cardinality0, cardinality1, cardinality2});
      outGroupIds = new int[numDocs];
      dictIds0 = new int[numDocs];
      dictIds1 = new int[numDocs];
      dictIds2 = new int[numDocs];
      for (int i = 0; i < numDocs; i++) {
        dictIds0[i] = ThreadLocalRandom.current().nextInt(cardinality0);
        dictIds1[i] = ThreadLocalRandom.current().nextInt(cardinality1);
        dictIds2[i] = ThreadLocalRandom.current().nextInt(cardinality2);
      }
      dictIds = new int[][] {dictIds0, dictIds1, dictIds2};
    }
  }

  public static class One {
    private int numKeys;
    private int groupCountUpperBound;
    private final boolean[] groups;

    public One(int cardinality) {
      this.groupCountUpperBound = cardinality;
      this.groups = new boolean[groupCountUpperBound];
    }
    public void apply(int numDocs, int[] dictIds, int[] outGroupIds) {
      System.arraycopy(dictIds, 0, outGroupIds, 0, numDocs);
      for (int i = 0; i < numDocs && numKeys < groupCountUpperBound; i++) {
        if (!groups[outGroupIds[i]]) {
          numKeys++;
          groups[outGroupIds[i]] = true;
        }
      }
    }
  }

  public static class Two {
    private final int groupCountUpperBound;
    private final boolean[] groups;
    private final int cardinality0;
    private int numKeys;


    public Two(int cardinality0, int cardinality1) {
      this.groupCountUpperBound = cardinality0 * cardinality1;
      this.groups = new boolean[groupCountUpperBound];
      this.cardinality0 = cardinality0;
    }

    public void apply(int numDocs, int[] dictIds0, int[] dictIds1, int[] outGroupIds) {
      for (int i = 0; i < numDocs; i++) {
        outGroupIds[i] = dictIds1[i] * cardinality0 + dictIds0[i];
      }
      for (int i = 0; i < numDocs && numKeys < groupCountUpperBound; i++) {
        if (!groups[outGroupIds[i]]) {
          numKeys++;
          groups[outGroupIds[i]] = true;
        }
      }
    }
  }

  public static class Three {
    private final int groupCountUpperBound;
    private final boolean[] groups;
    private final int cardinality0;
    private final int cardinality1;
    private int numKeys;


    public Three(int cardinality0, int cardinality1, int cardinality2) {
      this.groupCountUpperBound = cardinality0 * cardinality1 * cardinality2;
      this.groups = new boolean[groupCountUpperBound];
      this.cardinality0 = cardinality0;
      this.cardinality1 = cardinality1;
    }

    public void apply(int numDocs, int[] dictIds0, int[] dictIds1, int[] dictIds2, int[] outGroupIds) {
      for (int i = 0; i < numDocs; i++) {
        outGroupIds[i] = (dictIds2[i] * cardinality1 + dictIds1[i]) * cardinality0 + dictIds0[i];
      }
      for (int i = 0; i < numDocs && numKeys < groupCountUpperBound; i++) {
        if (!groups[outGroupIds[i]]) {
          numKeys++;
          groups[outGroupIds[i]] = true;
        }
      }
    }
  }

  public static class Generic {
    private final int groupCountUpperBound;
    private final int numExpressions;
    private final int[] cardinalities;
    private final boolean[] groups;
    int numKeys;

    public Generic(int[] cardinalities) {
      this.numExpressions = cardinalities.length;
      int count = 1;
      for (int cardinality : cardinalities) {
        count *= cardinality;
      }
      this.groupCountUpperBound = count;
      this.groups = new boolean[groupCountUpperBound];
      this.cardinalities = cardinalities;
    }

    public void apply(int numDocs, int[][] dictIds, int[] outGroupIds) {
      for (int i = 0; i < numDocs; i++) {
        int groupId = 0;
        for (int j = numExpressions - 1; j >= 0; j--) {
          groupId = groupId * cardinalities[j] + dictIds[j][i];
        }
        outGroupIds[i] = groupId;
        // if the flag is false, then increase the key num
        if (!groups[groupId]) {
          numKeys++;
        }
        groups[groupId] = true;
      }

    }
  }

  @Benchmark
  public int[] one(OneState state) {
    state.one.apply(state.numDocs, state.dictIds0, state.outGroupIds);
    return state.outGroupIds;
  }

  @Benchmark
  public int[] oneGeneric(OneState state) {
    state.generic.apply(state.numDocs, state.dictIds, state.outGroupIds);
    return state.outGroupIds;
  }

  @Benchmark
  public int[] two(TwoState state) {
    state.two.apply(state.numDocs, state.dictIds0, state.dictIds1, state.outGroupIds);
    return state.outGroupIds;
  }

  @Benchmark
  public int[] twoGeneric(TwoState state) {
    state.generic.apply(state.numDocs, state.dictIds, state.outGroupIds);
    return state.outGroupIds;
  }

  @Benchmark
  public int[] three(ThreeState state) {
    state.three.apply(state.numDocs, state.dictIds0, state.dictIds1, state.dictIds2, state.outGroupIds);
    return state.outGroupIds;
  }

  @Benchmark
  public int[] threeGeneric(ThreeState state) {
    state.generic.apply(state.numDocs, state.dictIds, state.outGroupIds);
    return state.outGroupIds;
  }
}
```

```
Benchmark                       (cardinality0)  (cardinality1)  (cardinality2)  (numDocs)  Mode  Cnt      Score       Error  Units
GroupByBenchmark.one                        10             N/A             N/A       1024  avgt    5     87.888 ±     0.035  ns/op
GroupByBenchmark.one                       100             N/A             N/A       1024  avgt    5     85.826 ±     0.090  ns/op
GroupByBenchmark.oneGeneric                 10             N/A             N/A       1024  avgt    5   3968.017 ±     4.523  ns/op
GroupByBenchmark.oneGeneric                100             N/A             N/A       1024  avgt    5   3967.884 ±     2.931  ns/op
GroupByBenchmark.three                      10               3               5       1024  avgt    5    180.661 ±     0.162  ns/op
GroupByBenchmark.three                      10               3               5      10000  avgt    5   2519.999 ±    15.816  ns/op
GroupByBenchmark.three                      10               3              50       1024  avgt    5    788.203 ±     0.850  ns/op
GroupByBenchmark.three                      10               3              50      10000  avgt    5   8369.957 ±    14.328  ns/op
GroupByBenchmark.three                      10              30               5       1024  avgt    5    789.272 ±     3.107  ns/op
GroupByBenchmark.three                      10              30               5      10000  avgt    5   8352.994 ±    25.510  ns/op
GroupByBenchmark.three                      10              30              50       1024  avgt    5    791.969 ±     0.755  ns/op
GroupByBenchmark.three                      10              30              50      10000  avgt    5   8456.677 ±    39.868  ns/op
GroupByBenchmark.three                     100               3               5       1024  avgt    5    788.635 ±     1.158  ns/op
GroupByBenchmark.three                     100               3               5      10000  avgt    5   8473.065 ±    43.604  ns/op
GroupByBenchmark.three                     100               3              50       1024  avgt    5    792.696 ±     1.293  ns/op
GroupByBenchmark.three                     100               3              50      10000  avgt    5   8400.931 ±    39.148  ns/op
GroupByBenchmark.three                     100              30               5       1024  avgt    5    792.267 ±     0.815  ns/op
GroupByBenchmark.three                     100              30               5      10000  avgt    5   8362.967 ±    46.329  ns/op
GroupByBenchmark.three                     100              30              50       1024  avgt    5   1009.376 ±     3.741  ns/op
GroupByBenchmark.three                     100              30              50      10000  avgt    5  11784.628 ±    75.056  ns/op
GroupByBenchmark.threeGeneric               10               3               5       1024  avgt    5   5168.556 ±    23.392  ns/op
GroupByBenchmark.threeGeneric               10               3               5      10000  avgt    5  55207.217 ± 26107.977  ns/op
GroupByBenchmark.threeGeneric               10               3              50       1024  avgt    5   5203.942 ±     7.881  ns/op
GroupByBenchmark.threeGeneric               10               3              50      10000  avgt    5  55742.896 ± 24792.893  ns/op
GroupByBenchmark.threeGeneric               10              30               5       1024  avgt    5   5163.989 ±     2.379  ns/op
GroupByBenchmark.threeGeneric               10              30               5      10000  avgt    5  55189.596 ± 26087.251  ns/op
GroupByBenchmark.threeGeneric               10              30              50       1024  avgt    5   5208.037 ±     9.829  ns/op
GroupByBenchmark.threeGeneric               10              30              50      10000  avgt    5  55353.802 ± 25864.541  ns/op
GroupByBenchmark.threeGeneric              100               3               5       1024  avgt    5   5163.058 ±     3.719  ns/op
GroupByBenchmark.threeGeneric              100               3               5      10000  avgt    5  55229.799 ± 26282.820  ns/op
GroupByBenchmark.threeGeneric              100               3              50       1024  avgt    5   5165.277 ±     1.592  ns/op
GroupByBenchmark.threeGeneric              100               3              50      10000  avgt    5  55250.657 ± 26038.710  ns/op
GroupByBenchmark.threeGeneric              100              30               5       1024  avgt    5   5248.230 ±    10.805  ns/op
GroupByBenchmark.threeGeneric              100              30               5      10000  avgt    5  55216.541 ± 26101.864  ns/op
GroupByBenchmark.threeGeneric              100              30              50       1024  avgt    5   5169.316 ±     7.386  ns/op
GroupByBenchmark.threeGeneric              100              30              50      10000  avgt    5  55558.478 ± 25691.232  ns/op
GroupByBenchmark.two                        10               3             N/A       1024  avgt    5    112.575 ±     0.119  ns/op
GroupByBenchmark.two                        10              30             N/A       1024  avgt    5    904.502 ±     0.849  ns/op
GroupByBenchmark.two                       100               3             N/A       1024  avgt    5    901.040 ±     3.153  ns/op
GroupByBenchmark.two                       100              30             N/A       1024  avgt    5    902.323 ±     0.614  ns/op
GroupByBenchmark.twoGeneric                 10               3             N/A       1024  avgt    5   4370.175 ±    21.262  ns/op
GroupByBenchmark.twoGeneric                 10              30             N/A       1024  avgt    5   4396.317 ±   129.928  ns/op
GroupByBenchmark.twoGeneric                100               3             N/A       1024  avgt    5   4372.822 ±    16.521  ns/op
GroupByBenchmark.twoGeneric                100              30             N/A       1024  avgt    5   4371.179 ±    21.749  ns/op
```

The difference in code generation can be seen with perfasm (for a 3D group by, cardinalities 100 x 30 x 50):

before:

```asm
....[Hottest Region 1]..............................................................................
c2, level 4, groupby.generated.GroupByBenchmark_threeGeneric_jmhTest::threeGeneric_avgt_jmhStub, version 497 (183 bytes) 

                 0x00007ff78b6eef95: lea    (%r12,%r11,8),%rbp
                 0x00007ff78b6eef99: vmovd  %r11d,%xmm1
                 0x00007ff78b6eef9e: vmovd  %xmm2,%r11d
                 0x00007ff78b6eefa3: lea    (%r12,%r11,8),%rax
                 0x00007ff78b6eefa7: xor    %ecx,%ecx
         ╭       0x00007ff78b6eefa9: jmp    0x00007ff78b6eefdf
         │↗      0x00007ff78b6eefab: xor    %r8d,%r8d          ;*iflt {reexecute=0 rethrow=0 return_oop=0}
         ││                                                    ; - groupby.GroupByBenchmark$Generic::apply@22 (line 213)
         ││                                                    ; - groupby.GroupByBenchmark::threeGeneric@16 (line 300)
         ││                                                    ; - groupby.generated.GroupByBenchmark_threeGeneric_jmhTest::threeGeneric_avgt_jmhStub@19 (line 237)
  5.10%  ││   ↗  0x00007ff78b6eefae: mov    %r8d,0x10(%rbp,%rcx,4)  ;*iastore {reexecute=0 rethrow=0 return_oop=0}
         ││   │                                                ; - groupby.GroupByBenchmark$Generic::apply@56 (line 216)
         ││   │                                                ; - groupby.GroupByBenchmark::threeGeneric@16 (line 300)
         ││   │                                                ; - groupby.generated.GroupByBenchmark_threeGeneric_jmhTest::threeGeneric_avgt_jmhStub@19 (line 237)
  0.73%  ││   │  0x00007ff78b6eefb3: cmp    %r14d,%r8d
         ││   │  0x00007ff78b6eefb6: jae    0x00007ff78b6ef0d0
  0.09%  ││   │  0x00007ff78b6eefbc: movslq %r8d,%r11
  1.00%  ││   │  0x00007ff78b6eefbf: movzbl 0x10(%rax,%r11,1),%edx  ;*baload {reexecute=0 rethrow=0 return_oop=0}
         ││   │                                                ; - groupby.GroupByBenchmark$Generic::apply@63 (line 218)
         ││   │                                                ; - groupby.GroupByBenchmark::threeGeneric@16 (line 300)
         ││   │                                                ; - groupby.generated.GroupByBenchmark_threeGeneric_jmhTest::threeGeneric_avgt_jmhStub@19 (line 237)
 28.57%  ││   │  0x00007ff78b6eefc5: test   %edx,%edx
         ││   │  0x00007ff78b6eefc7: je     0x00007ff78b6ef13c  ;*ifne {reexecute=0 rethrow=0 return_oop=0}
         ││   │                                                ; - groupby.GroupByBenchmark$Generic::apply@64 (line 218)
         ││   │                                                ; - groupby.GroupByBenchmark::threeGeneric@16 (line 300)
         ││   │                                                ; - groupby.generated.GroupByBenchmark_threeGeneric_jmhTest::threeGeneric_avgt_jmhStub@19 (line 237)
  5.57%  ││   │  0x00007ff78b6eefcd: movb   $0x1,0x10(%rax,%r11,1)  ;*bastore {reexecute=0 rethrow=0 return_oop=0}
         ││   │                                                ; - groupby.GroupByBenchmark$Generic::apply@84 (line 221)
         ││   │                                                ; - groupby.GroupByBenchmark::threeGeneric@16 (line 300)
         ││   │                                                ; - groupby.generated.GroupByBenchmark_threeGeneric_jmhTest::threeGeneric_avgt_jmhStub@19 (line 237)
  0.22%  ││   │  0x00007ff78b6eefd3: inc    %ecx               ;*iinc {reexecute=0 rethrow=0 return_oop=0}
         ││   │                                                ; - groupby.GroupByBenchmark$Generic::apply@85 (line 211)
         ││   │                                                ; - groupby.GroupByBenchmark::threeGeneric@16 (line 300)
         ││   │                                                ; - groupby.generated.GroupByBenchmark_threeGeneric_jmhTest::threeGeneric_avgt_jmhStub@19 (line 237)
  0.31%  ││   │  0x00007ff78b6eefd5: cmp    0x70(%rsp),%ecx
  0.02%  ││   │  0x00007ff78b6eefd9: jge    0x00007ff78b6eeeab  ;*iconst_0 {reexecute=0 rethrow=0 return_oop=0}
         ││   │                                                ; - groupby.GroupByBenchmark$Generic::apply@9 (line 212)
         ││   │                                                ; - groupby.GroupByBenchmark::threeGeneric@16 (line 300)
         ││   │                                                ; - groupby.generated.GroupByBenchmark_threeGeneric_jmhTest::threeGeneric_avgt_jmhStub@19 (line 237)
  1.51%  ↘│   │  0x00007ff78b6eefdf: test   %r10d,%r10d
          ╰   │  0x00007ff78b6eefe2: jl     0x00007ff78b6eefab  ;*iflt {reexecute=0 rethrow=0 return_oop=0}
              │                                                ; - groupby.GroupByBenchmark$Generic::apply@22 (line 213)
              │                                                ; - groupby.GroupByBenchmark::threeGeneric@16 (line 300)
              │                                                ; - groupby.generated.GroupByBenchmark_threeGeneric_jmhTest::threeGeneric_avgt_jmhStub@19 (line 237)
  5.23%       │  0x00007ff78b6eefe4: mov    %r10d,%edx
  0.40%       │  0x00007ff78b6eefe7: xor    %r8d,%r8d          ;*iload {reexecute=0 rethrow=0 return_oop=0}
              │                                                ; - groupby.GroupByBenchmark$Generic::apply@25 (line 214)
              │                                                ; - groupby.GroupByBenchmark::threeGeneric@16 (line 300)
              │                                                ; - groupby.generated.GroupByBenchmark_threeGeneric_jmhTest::threeGeneric_avgt_jmhStub@19 (line 237)
  0.20%    ↗  │  0x00007ff78b6eefea: mov    0x10(%r9,%rdx,4),%esi  ;*aaload {reexecute=0 rethrow=0 return_oop=0}
           │  │                                                ; - groupby.GroupByBenchmark$Generic::apply@38 (line 214)
…