@richardstartin 2021-12-22
@kishoreg it's still a draft, I know that the 2D special case is integer multiples faster, but I made a mistake with the 3D case and need to reevaluate.