@siddharthteotia 2021-12-23
I am guessing the `vpaddd` and other similar instructions are the SIMD instructions getting used in the second region in the JIT compiled code and that sort of proves that new code is getting auto-vectorized. Is this correct ?