@richardstartin 2021-12-23
Yes, the multiplications (`vpmulld`) and additions (`vpaddd`) are vectorised so 8 are performed at a time after the transformation, which explains the speedup.